Use float value for a factor in depth integrator

+++ src/integrators/depth.h
@@ -45,7 +45,7 @@ public:
     void RequestSamples(Sampler *sampler, Sample *sample, const Scene *scene);
 private:
     // DepthSurfaceIntegrator Private Data
-    int factor;
+    float factor;
 };
